Fullstack/Backend/requirements.txt
asgiref==3.8.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.32.3
SQLAlchemy==2.0.38
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==2.3.7
